split object into a front-end and detail, to fix unresolved ref
