+++ luabind/detail/constructor.hpp
@@ -8,7 +8,7 @@
 #  define LUABIND_DETAIL_CONSTRUCTOR_081018_HPP
 
 #  include <luabind/get_main_thread.hpp>
-#  include <luabind/object.hpp>
+#  include <luabind/detail/object.hpp>
 #  include <luabind/wrapper_base.hpp>
 #  include <luabind/detail/inheritance.hpp>
